+++ core/src/server/delivery/request/GetVectorByIDRequest.cpp
@@ -76,7 +76,12 @@ GetVectorByIDRequest::OnExecute() {
         }
 
         // step 2: get vector data, now only support get one id
-        return DBWrapper::DB()->GetVectorByID(collection_name_, ids_[0], vectors_);
+        std::vector<engine::VectorsData> vectors;
+        status = DBWrapper::DB()->GetVectorsByID(collection_name_, ids_, vectors);
+        if (!vectors.empty()) {
+            vectors_ = vectors[0];
+        }
+        return status;
     } catch (std::exception& ex) {
         return Status(SERVER_UNEXPECTED_ERROR, ex.what());
     }